Release 1.2.0

+++ test/libec_slap.c
@@ -42,6 +42,7 @@
 #include <assert.h>
 #include "erasurecode.h"
 #include "erasurecode_helpers.h"
+#include "erasurecode_helpers_ext.h"
 #include "builtin/xor_codes/test_xor_hd_code.h"
 
 struct frag_array_set {
@@ -52,7 +53,7 @@ struct frag_array_set {
 void print_mask(unsigned long mask)
 {
     unsigned int i = 0;
-    unsigned long pos = 1;
+    long pos = 1;
 
     if (mask == 0) {
         fprintf(stderr,"  No Missing fragments");
@@ -69,7 +70,7 @@ void print_mask(unsigned long mask)
 void missing_mask_to_array(long mask, int *missing)
 {
     unsigned int i = 0;
-    unsigned long pos = 1;
+    long pos = 1;
 
     for (i = 0; i < (sizeof(size_t) * 8) - 1; i++) {
         if ((mask & (pos << i)) != 0) {
@@ -78,7 +79,7 @@ void missing_mask_to_array(long mask, int *missing)
     }
 }
 
-size_t add_item_to_missing_mask(unsigned long mask, unsigned int pos) 
+size_t add_item_to_missing_mask(unsigned long mask, long pos) 
 {
     if (pos < 0) {
         return mask;
